Mover verify_certificate to Certificate class
